single-quote frame

+++ lua/packages/which-key.lua
@@ -2,7 +2,7 @@ local wk = require('which-key')
 
 local setup = {
   window = {
-    border = "double", -- none, single, double, shadow
+    border = "single", -- none, single, double, shadow
     position = "bottom", -- bottom, top
     margin = { 1, 0, 1, 0 }, -- extra window margin [top, right, bottom, left]
     padding = { 2, 2, 2, 2 }, -- extra window padding [top, right, bottom, left]
